# Improve performance of RB circuit generation